:wrench: Refactored CreateOrSelect and extracted form steps
