+++ src/Components/Forms/Form.scss
@@ -34,3 +34,14 @@
     }
   }
 }
+
+.select-or-create-field {
+  display: flex;
+  flex-direction: row-reverse;
+  width: 100%;
+  margin-bottom: 20px;
+
+  .ant-btn {
+    width: 50%;
+  }
+}
