+++ src/Components/Upload/index.js
@@ -1,14 +1,10 @@
 import React from 'react'
-import { ExperimentOutlined, FileAddOutlined } from '@ant-design/icons'
-import { Typography, message, Modal } from 'antd'
-import RadioGroup from '~/src/Components/Forms/RadioGroup'
-import TextArea from '~/src/Components/Forms/TextArea'
+import { message, Modal } from 'antd'
 import SteppedFormFactory from '../Forms/SteppedForm'
-import FileInput from '../Forms/FileInput'
-import CreateStudy from './CreateStudy'
 import { useHistory } from 'react-router-dom'
 import apiWebSocket from '../../api/apiWebSocket'
 import ExtraInfo from './ExtraInfo'
+import uploadSteps from './uploadForm'
 
 const Upload = () => {
   const [loading, setLoading] = React.useState(false)
@@ -20,76 +16,6 @@ const Upload = () => {
   const [connectionSocket, setConnectionSocket] = React.useState(null)
 
   const history = useHistory()
-
-  const steps = [
-    {
-      title: 'Upload File',
-      icon: <FileAddOutlined />,
-      fields: [
-        {
-          name: 'study',
-          label: 'Study Name',
-          showLabel: true,
-          RenderField: CreateStudy,
-          rules: [{ required: true }],
-        },
-        {
-          name: 'machine',
-          label: 'Machine',
-          showLabel: true,
-          placeholder: 'Machine',
-          options: ['HTX Synergy', 'BMG', 'FluoPi'],
-          RenderField: RadioGroup,
-          rules: [{ required: true }],
-        },
-        {
-          name: 'data_file',
-          label: 'Data File',
-          showLabel: true,
-          RenderField: FileInput,
-          rules: [{ required: true }],
-        },
-      ],
-    },
-    {
-      title: 'Assay',
-      icon: <ExperimentOutlined />,
-      fields: [
-        {
-          name: 'mock_title2',
-          label: 'mock_title2',
-          RenderField() {
-            return (
-              <Typography.Title level={3}>
-                Assays description and temperature
-              </Typography.Title>
-            )
-          },
-        },
-        {
-          name: 'name',
-          label: 'Name',
-          showLabel: true,
-          rules: [{ required: true, max: 80, min: 3 }],
-        },
-        {
-          name: 'description',
-          label: 'Description',
-          showLabel: true,
-          RenderField: TextArea,
-          rules: [{ required: true }],
-        },
-        {
-          name: 'temperature',
-          label: 'Temperature',
-          showLabel: true,
-          type: 'number',
-          rules: [{ required: true }],
-          addonAfter: '°C',
-        },
-      ],
-    },
-  ]
 
   const onSubmit = async (data) => {
     const { name, machine, description, temperature, study } = data
@@ -156,7 +82,7 @@ const Upload = () => {
     <>
       <SteppedFormFactory
         name="UpladForm"
-        steps={steps}
+        steps={uploadSteps}
         onSubmit={onSubmit}
         submitText="Submit"
         loading={loading}
